requirements.txt
pymongo==4.13.2
python-telegram-bot==22.1